+++ common/src/main/java/dev/terminalmc/modlistmemory/ModListMemory.java
@@ -19,9 +19,9 @@ package dev.terminalmc.modlistmemory;
 import dev.terminalmc.modlistmemory.config.Config;
 import dev.terminalmc.modlistmemory.util.ModLogger;
 import net.minecraft.client.Minecraft;
-
-import java.util.ArrayList;
-import java.util.List;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.util.FormattedCharSequence;
 
 import static dev.terminalmc.modlistmemory.config.Config.options;
 
@@ -29,43 +29,75 @@ public class ModListMemory {
     public static final String MOD_ID = "modlistmemory";
     public static final String MOD_NAME = "ModListMemory";
     public static final ModLogger LOG = new ModLogger(MOD_NAME);
+    public static final int PINNED_BADGE_OUTLINE = 0xFFb36b19;
+    public static final int PINNED_BADGE_FILL = 0xFF4d2e0b;
+    public static final int RECENT_BADGE_OUTLINE = 0xFF1d9fb3;
+    public static final int RECENT_BADGE_FILL = 0xFF0c444d;
 
-    public static List<String> recentMods = new ArrayList<>();
-    public static double scrollAmount = Double.MIN_VALUE;
+    public static FormattedCharSequence pinnedBadgeText = Component.literal("I").getVisualOrderText();
+    public static FormattedCharSequence recentBadgeText = Component.literal("R").getVisualOrderText();
 
     public static void init() {
         Config.getAndSave();
         
-        if (options().persistOnRestart) {
-            recentMods.addAll(options().recentMods);
-            scrollAmount = options().scrollAmount;
+        if (!options().persistOnRestart) {
+            options().recentMods.clear();
+            options().scrollAmount = Config.Options.scrollAmountDefault;
         }
     }
 
     public static void onClientShutdown(Minecraft mc) {
         if (options().persistOnRestart) {
-            saveMemoryToConfig();
+            Config.save();
         }
     }
     
     public static void onConfigSaved(Config config) {
-        // If you are maintaining caches based on config values, update them here.
+        pinnedBadgeText = Component.literal(config.options.pinnedText).getVisualOrderText();
+        recentBadgeText = Component.literal(config.options.recentText).getVisualOrderText();
     }
 
     public static void onModOpened(String modId) {
-        recentMods.remove(modId);
-        recentMods.addLast(modId);
-        while (recentMods.size() > options().memorySize) {
-            recentMods.removeFirst();
+        if (options().pinnedMods.contains(modId)) return; // Ignore pinned mods
+        // Move to the top of the recent list
+        options().recentMods.remove(modId);
+        options().recentMods.addLast(modId);
+        while (options().recentMods.size() > options().recentModsSize) {
+            options().recentMods.removeFirst();
         }
         if (options().persistOnRestart && options().saveOnUpdate) {
-            saveMemoryToConfig();
+            Config.save();
         }
     }
     
-    public static void saveMemoryToConfig() {
-        options().recentMods = recentMods;
-        options().scrollAmount = scrollAmount;
-        Config.save();
+    public static void onModClicked(String modId) {
+        if (ModListMemory.hasKeyDown(options().pinKey)) {
+            // Pin, or move to the top of the pin list
+            options().recentMods.remove(modId);
+            options().pinnedMods.remove(modId);
+            options().pinnedMods.addLast(modId);
+            while (options().pinnedMods.size() > options().pinnedModsSize) {
+                options().pinnedMods.removeFirst();
+            }
+            if (options().persistOnRestart && options().saveOnUpdate) {
+                Config.save();
+            }
+        } 
+        else if (ModListMemory.hasKeyDown(options().unpinKey)) {
+            // Unpin
+            options().pinnedMods.remove(modId);
+            if (options().persistOnRestart && options().saveOnUpdate) {
+                Config.save();
+            }
+        }
+    }
+    
+    public static boolean hasKeyDown(Config.Key key) {
+        return switch(key) {
+            case CONTROL -> Screen.hasControlDown();
+            case ALT -> Screen.hasAltDown();
+            case SHIFT -> Screen.hasShiftDown();
+            case NONE -> false;
+        };
     }
 }