+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Added ability to pin and unpin mods using shift and control click
+- Added badges for pinned and recent mods
+
 ## 0.1.1
 
 - Fixed version metadata